requirements.txt
python-telegram-bot==21.5
openai==1.3.9
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
